Split the execution engine into soldb-evm
soldb-rpc held two things: the JSON-RPC transport and everything that
turns a node's answers into a trace or re-executes one in REVM. A
WebAssembly host has no socket and supplies the answers itself, and
`soldb run` involves no node at all, yet both had to depend on the
transport to reach the engine.

The engine now lives in soldb-evm: the node data shapes, the trace
assembly every backend ends in, and the REVM replay behind the same
default-on `replay` feature, including the state providers a host fills
and the local chain `run` executes on. Nothing in it does I/O. soldb-rpc
keeps the client, the lookups, the `debug-rpc` backend, and the node
side of replay: fetching a transaction's inputs, the lazy state provider
with its archive preflight, and batch requests. It re-exports the moved
surface, so no dependent changes.

The move is otherwise verbatim. ReplayInputs::fetch became a function on
the node side, with the block-number check it used made public; the
quantity and address helpers the node side calls, and the account
builder, are public; the replay error's message is a public field so the
provider can construct it; and the two provider tests read the provider
through its own trait instead of the engine's private adapter, which
forwards to it unchanged. The primitive types the provider trait names
are re-exported so an implementor needs no REVM dependency of its own.

The release workflow publishes twelve crates now; crates.io trusted
publishing has to be configured for soldb-evm before the next release.
