+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "crates/soldb-debugger",
     "crates/soldb-dap",
     "crates/soldb-ethdebug",
+    "crates/soldb-evm",
     "crates/soldb-profiler",
     "crates/soldb-repl",
     "crates/soldb-rpc",
@@ -36,6 +37,9 @@ soldb-core = { path = "crates/soldb-core", version = "0.2.0" }
 soldb-dap = { path = "crates/soldb-dap", version = "0.2.0" }
 soldb-debugger = { path = "crates/soldb-debugger", version = "0.2.0" }
 soldb-ethdebug = { path = "crates/soldb-ethdebug", version = "0.2.0" }
+# Off by default here so a crate that wants the transport without REVM can say so; the
+# `replay` feature of each dependent turns it back on.
+soldb-evm = { path = "crates/soldb-evm", version = "0.2.0", default-features = false }
 soldb-profiler = { path = "crates/soldb-profiler", version = "0.2.0" }
 soldb-repl = { path = "crates/soldb-repl", version = "0.2.0" }
 soldb-rpc = { path = "crates/soldb-rpc", version = "0.2.0" }
