+++ crates/soldb-rpc/Cargo.toml
@@ -2,7 +2,7 @@
 name = "soldb-rpc"
 version.workspace = true
 edition.workspace = true
-description = "JSON-RPC and REVM replay backends that produce SolDB transaction traces"
+description = "JSON-RPC transport and the node-backed trace backends for SolDB"
 license.workspace = true
 repository.workspace = true
 homepage.workspace = true
@@ -15,32 +15,19 @@ doctest = false
 
 [features]
 default = ["replay"]
-# The REVM replay backend. It is the only part of the crate that links REVM, so a build
-# without it, such as the WebAssembly package, keeps the JSON-RPC transport, the
-# `debug-rpc` backend, and trace assembly at a fraction of the size.
-replay = ["dep:revm", "dep:getrandom"]
+# The REVM replay backend: the engine in `soldb-evm` plus the node side here that fetches
+# a transaction's inputs and reads parent-block state over RPC. Off, a build such as the
+# WebAssembly package keeps the JSON-RPC transport and the `debug-rpc` backend without
+# linking REVM.
+replay = ["soldb-evm/replay"]
 
 [dependencies]
-# `U256` for wei quantities. This is the same `ruint` type `revm::primitives::U256`
-# re-exports, so the replay backend and the shared quantity helpers never convert.
-ruint = { version = "1.18", default-features = false }
 serde.workspace = true
 serde_json.workspace = true
 soldb-core.workspace = true
-revm = { version = "36.0", optional = true, default-features = false, features = [
-    "std",
-    "serde",
-    "optional_eip3607",
-    "optional_block_gas_limit",
-    "optional_no_base_fee",
-] }
-
-# `revm` reaches `getrandom` through `k256`, and `getrandom` 0.2 refuses to build for
-# `wasm32-unknown-unknown` unless its `js` backend is selected. Nothing in this crate
-# draws randomness; the dependency exists only to pick that backend so the REVM replay
-# path links in a browser or Node.js host when the `replay` feature is on there.
-[target.'cfg(all(target_arch = "wasm32", target_os = "unknown"))'.dependencies]
-getrandom = { version = "0.2", features = ["js"], optional = true }
+# The workspace entry turns the engine's default features off, so its `replay` feature
+# follows ours instead of turning itself on underneath a transport-only build.
+soldb-evm.workspace = true
 
 [lints]
 workspace = true
